// models/BookInterface.tsx
export default interface BookInterface {
  _id: string;
  title: string;
  author: string;
  price: number;
  trending: boolean;
  newRelease: boolean;
  categories: string[];
  img: string;
  description: string;
  publicationDate: string;
  ISBN: string;
  dimension: string;
  weight: number;
  language: string;
  publisher: string;
}